+++ apps/router-e2e/__e2e__/01-rsc/app/index.tsx
@@ -1,4 +1,4 @@
-import { Text, View } from '../lib/react-native';
+import { Image, Text, View } from '../lib/react-native';
 
 import { Counter } from '../components/counter';
 
@@ -9,6 +9,12 @@ export default function IndexRoute() {
       <Text>Platform: {process.env.EXPO_OS}</Text>
       <Text testID="secret-text">Secret: {process.env.TEST_SECRET_VALUE}</Text>
       <Text>Render: {Date.now()}</Text>
+
+      <Image
+        testID="main-image"
+        source={require('../../../assets/icon.png')}
+        style={{ width: 100, height: 100 }}
+      />
 
       <Counter />
     </View>