+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -640,12 +640,13 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     files: ExportAssetMap;
   }> {
     // NOTE(EvanBacon): This will not support any code elimination since it's a static pass.
-    const clientBoundaries = await this.rscRenderer!.getExpoRouterClientReferencesAsync(
-      {
-        platform: options.platform,
-      },
-      files
-    );
+    const { reactClientReferences: clientBoundaries, cssModules } =
+      await this.rscRenderer!.getExpoRouterClientReferencesAsync(
+        {
+          platform: options.platform,
+        },
+        files
+      );
 
     debug('Evaluated client boundaries:', clientBoundaries);
 
@@ -657,6 +658,9 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       },
       extraOptions
     );
+
+    // Inject the global CSS that was imported during the server render.
+    bundle.artifacts.push(...cssModules);
 
     const serverRoot = getMetroServerRoot(this.projectRoot);
 
@@ -935,6 +939,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
 
       // If React 19 is enabled, then add RSC middleware to the dev server.
       if (isReactServerComponentsEnabled) {
+        this.bindRSCDevModuleInjectionHandler();
         const rscMiddleware = createServerComponentsMiddleware(this.projectRoot, {
           instanceMetroOptions: this.instanceMetroOptions,
           rscPath: '/_flight',
@@ -1002,6 +1007,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     } else {
       // If React 19 is enabled, then add RSC middleware to the dev server.
       if (isReactServerComponentsEnabled) {
+        this.bindRSCDevModuleInjectionHandler();
         const rscMiddleware = createServerComponentsMiddleware(this.projectRoot, {
           instanceMetroOptions: this.instanceMetroOptions,
           rscPath: '/_flight',
@@ -1258,6 +1264,25 @@ export class MetroBundlerDevServer extends BundlerDevServer {
 
   private invalidateApiRouteCache() {
     this.pendingRouteOperations.clear();
+  }
+
+  // Ensure the global is available for SSR CSS modules to inject client updates.
+  private bindRSCDevModuleInjectionHandler() {
+    // Used by SSR CSS modules to broadcast client updates.
+    // @ts-expect-error
+    globalThis.__expo_rsc_inject_module = this.sendClientModule.bind(this);
+  }
+
+  // NOTE: This can only target a single platform at a time (web).
+  // used for sending RSC CSS to the root client in development.
+  private sendClientModule({ code, id }: { code: string; id: string }) {
+    this.broadcastMessage('sendDevCommand', {
+      name: 'module-import',
+      data: {
+        code,
+        id,
+      },
+    });
   }
 
   // Metro HMR